Moved 15 funcs to `text_replace.py`: pronoun_repl, name_repl, process_text, adjust_prey_abbr, get_special_snippet_list, find_special_list_types, ongoing_event_text_adjust, event_text_adjust, leader_ceremony_text_adjust, ceremony_text_adjust, get_leader_life_notice, adjust_list_text, history_text_adjust, selective_replace, shorten_text_to_fit
