+++ scripts/events_module/patrol/patrol_outcome.py
@@ -17,10 +17,7 @@ from scripts.game_structure.game.settings import game_setting_get
 if TYPE_CHECKING:
     from scripts.events_module.patrol.patrol import Patrol
 
-from scripts.utility import (
-    event_text_adjust,
-    adjust_list_text,
-)
+from scripts.events_module.text_adjust import event_text_adjust, adjust_list_text
 from scripts.events_module.consequences import (
     create_new_cat_block,
     gather_cat_objects,
