+++ scripts/events_module/relationship/group_events.py
@@ -20,9 +20,7 @@ from scripts.events_module.event_filters import (
     event_for_cat,
     filter_relationship_type,
 )
-from scripts.utility import (
-    process_text,
-)
+from scripts.events_module.text_adjust import process_text
 from scripts.events_module.consequences import change_relationship_values
 from scripts.game_structure.localization import load_lang_resource
