+++ scripts/events_module/relationship/romantic_events.py
@@ -13,10 +13,7 @@ from scripts.event_class import Single_Event
 from scripts.game_structure import constants
 from scripts.game_structure import game
 from scripts.game_structure.localization import load_lang_resource
-from scripts.utility import (
-    event_text_adjust,
-    process_text,
-)
+from scripts.events_module.text_adjust import process_text, event_text_adjust
 from scripts.events_module.consequences import change_relationship_values
 from scripts.events_module.event_filters import (
     get_highest_romantic_relation,
