+++ scripts/events_module/short/condition_events.py
@@ -30,10 +30,7 @@ from scripts.game_structure.game.switches import (
 )
 from scripts.game_structure import game
 from scripts.game_structure.localization import load_lang_resource
-from scripts.utility import (
-    event_text_adjust,
-    get_leader_life_notice,
-)
+from scripts.events_module.text_adjust import event_text_adjust, get_leader_life_notice
 from scripts.clan_package.get_clan_cats import find_alive_cats_with_rank
