+++ scripts/screens/ChangeGenderScreen.py
@@ -18,8 +18,8 @@ from scripts.game_structure.ui_elements import (
 )
 from scripts.utility import (
     get_text_box_theme,
-    shorten_text_to_fit,
 )
+from ..events_module.text_adjust import shorten_text_to_fit
 from ..cat import pronouns
 from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset, ui_scale_value
 from .Screens import Screens
