+++ scripts/screens/ChooseMateScreen.py
@@ -13,8 +13,8 @@ from scripts.game_structure.ui_elements import (
 )
 from scripts.utility import (
     get_text_box_theme,
-    shorten_text_to_fit,
 )
+from ..events_module.text_adjust import shorten_text_to_fit
 from ..events_module.event_filters import get_personality_compatibility
 from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset
 from .Screens import Screens
