+++ scripts/screens/FamilyTreeScreen.py
@@ -13,9 +13,8 @@ from scripts.game_structure.ui_elements import (
 )
 from scripts.utility import (
     get_text_box_theme,
-    shorten_text_to_fit,
-    adjust_list_text,
 )
+from ..events_module.text_adjust import adjust_list_text, shorten_text_to_fit
 from ..ui.scale import ui_scale, ui_scale_dimensions
 from .Screens import Screens
 from .enums import GameScreen
