+++ scripts/screens/SpriteInspectScreen.py
@@ -9,9 +9,9 @@ from scripts.game_structure import game
 from scripts.game_structure.ui_elements import UIImageButton, UISurfaceImageButton
 from scripts.utility import (
     generate_sprite,
-    shorten_text_to_fit,
     get_text_box_theme,
 )
+from ..events_module.text_adjust import shorten_text_to_fit
 from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset
 from .Screens import Screens
 from .enums import GameScreen
