+++ tests/test_lang.py
@@ -13,7 +13,7 @@ from scripts.game_structure.localization import (
     set_lang_config_directory,
 )
 from scripts.cat.pronouns import get_new_pronouns, determine_plural_pronouns
-from scripts.utility import event_text_adjust
+from scripts.events_module.text_adjust import event_text_adjust
 
 
 class TestLocalisation(unittest.TestCase):
