+++ tests/test_pronouns.py
@@ -19,7 +19,7 @@ os.environ["SDL_VIDEODRIVER"] = "dummy"
 os.environ["SDL_AUDIODRIVER"] = "dummy"
 
 from scripts.cat.pronouns import get_new_pronouns
-from scripts.utility import process_text
+from scripts.events_module.text_adjust import process_text
 
 
 def _test():
